backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
yfinance==0.2.36
pandas==2.2.0
requests==2.31.0
python-multipart==0.0.6
